@@ -39,9 +39,10 @@
     "karma-webpack": "^1.7.0",
     "mocha": "^2.3.3",
     "phantomjs": "^1.9.18",
-    "react": "^15.0.1",
-    "react-addons-test-utils": "^15.0.1",
-    "react-dom": "^15.0.1",
+    "prop-types": "^15.5.8",
+    "react": "^15.3.0",
+    "react-addons-test-utils": "^15.3.0",
+    "react-dom": "^15.3.0",
     "react-hot-loader": "^1.3.0",
     "sinon": "^1.17.1",
     "sinon-chai": "^2.8.0",
@@ -51,8 +52,9 @@
     "webpack-dev-server": "^1.12.0"
   },
   "peerDependencies": {
-    "react": "^0.14.8 || ^15.0.1",
-    "react-dom": "^0.14.8 || ^15.0.1"
+    "prop-types": "^15.5.8",
+    "react": "^0.14.9 || ^15.3.0",
+    "react-dom": "^0.14.9 || ^15.3.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
