@@ -16,6 +16,10 @@ var _reactDom = require('react-dom');
 
 var _reactDom2 = _interopRequireDefault(_reactDom);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _kwReactTweenState = require('kw-react-tween-state');
 
 var _kwReactTweenState2 = _interopRequireDefault(_kwReactTweenState);
@@ -64,34 +68,34 @@ var Carousel = _react2['default'].createClass({
   mixins: [_kwReactTweenState2['default'].Mixin],
 
   propTypes: {
-    afterSlide: _react2['default'].PropTypes.func,
-    autoplay: _react2['default'].PropTypes.bool,
-    autoplayInterval: _react2['default'].PropTypes.number,
-    beforeSlide: _react2['default'].PropTypes.func,
-    cellAlign: _react2['default'].PropTypes.oneOf(['left', 'center', 'right']),
-    cellSpacing: _react2['default'].PropTypes.number,
-    data: _react2['default'].PropTypes.func,
-    decorators: _react2['default'].PropTypes.arrayOf(_react2['default'].PropTypes.shape({
-      component: _react2['default'].PropTypes.func,
-      position: _react2['default'].PropTypes.oneOf(['TopLeft', 'TopCenter', 'TopRight', 'CenterLeft', 'CenterCenter', 'CenterRight', 'BottomLeft', 'BottomCenter', 'BottomRight']),
-      style: _react2['default'].PropTypes.object
+    afterSlide: _propTypes2['default'].func,
+    autoplay: _propTypes2['default'].bool,
+    autoplayInterval: _propTypes2['default'].number,
+    beforeSlide: _propTypes2['default'].func,
+    cellAlign: _propTypes2['default'].oneOf(['left', 'center', 'right']),
+    cellSpacing: _propTypes2['default'].number,
+    data: _propTypes2['default'].func,
+    decorators: _propTypes2['default'].arrayOf(_propTypes2['default'].shape({
+      component: _propTypes2['default'].func,
+      position: _propTypes2['default'].oneOf(['TopLeft', 'TopCenter', 'TopRight', 'CenterLeft', 'CenterCenter', 'CenterRight', 'BottomLeft', 'BottomCenter', 'BottomRight']),
+      style: _propTypes2['default'].object
     })),
-    dragging: _react2['default'].PropTypes.bool,
-    easing: _react2['default'].PropTypes.string,
-    edgeEasing: _react2['default'].PropTypes.string,
-    framePadding: _react2['default'].PropTypes.string,
-    frameOverflow: _react2['default'].PropTypes.string,
-    initialSlideHeight: _react2['default'].PropTypes.number,
-    initialSlideWidth: _react2['default'].PropTypes.number,
-    slideIndex: _react2['default'].PropTypes.number,
-    slidesToShow: _react2['default'].PropTypes.number,
-    slidesToScroll: _react2['default'].PropTypes.oneOfType([_react2['default'].PropTypes.number, _react2['default'].PropTypes.oneOf(['auto'])]),
-    slideWidth: _react2['default'].PropTypes.oneOfType([_react2['default'].PropTypes.string, _react2['default'].PropTypes.number]),
-    speed: _react2['default'].PropTypes.number,
-    swiping: _react2['default'].PropTypes.bool,
-    vertical: _react2['default'].PropTypes.bool,
-    width: _react2['default'].PropTypes.string,
-    wrapAround: _react2['default'].PropTypes.bool
+    dragging: _propTypes2['default'].bool,
+    easing: _propTypes2['default'].string,
+    edgeEasing: _propTypes2['default'].string,
+    framePadding: _propTypes2['default'].string,
+    frameOverflow: _propTypes2['default'].string,
+    initialSlideHeight: _propTypes2['default'].number,
+    initialSlideWidth: _propTypes2['default'].number,
+    slideIndex: _propTypes2['default'].number,
+    slidesToShow: _propTypes2['default'].number,
+    slidesToScroll: _propTypes2['default'].oneOfType([_propTypes2['default'].number, _propTypes2['default'].oneOf(['auto'])]),
+    slideWidth: _propTypes2['default'].oneOfType([_propTypes2['default'].string, _propTypes2['default'].number]),
+    speed: _propTypes2['default'].number,
+    swiping: _propTypes2['default'].bool,
+    vertical: _propTypes2['default'].bool,
+    width: _propTypes2['default'].string,
+    wrapAround: _propTypes2['default'].bool
   },
 
   getDefaultProps: function getDefaultProps() {
